Move `DispatchQueueDecorator` to seperate file
