+++ EssentialFeediOS/Feed UI/Composer/DispatchQueueDecorator.swift
@@ -1,0 +1,41 @@
+//
+//  DispatchQueueDecorator.swift
+//  EssentialFeediOS
+//
+//  Created by Kumar, Sanjay (623) on 16/06/23.
+//
+import Foundation
+import EssentialFeed
+
+final class DispatchQueueDecorator<T> {
+    private let decoratee: T
+    init(decoratee: T) {
+        self.decoratee = decoratee
+    }
+    private func dispatch(_ completion: @escaping () -> Void) {
+        guard Thread.isMainThread else {
+            return DispatchQueue.main.async(execute: completion)
+        }
+        completion()
+    }
+}
+
+extension DispatchQueueDecorator: FeedLoader where T == FeedLoader {
+    func load(completion: @escaping (FeedLoader.Result) -> Void) {
+        decoratee.load { [weak self] result in
+            self?.dispatch {
+                completion(result)
+            }
+        }
+    }
+}
+
+extension DispatchQueueDecorator: FeedImageDataLoader where T == FeedImageDataLoader {
+    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
+        return decoratee.loadImageData(from: url) { [weak self] result in
+            self?.dispatch {
+                completion(result)
+            }
+        }
+    }
+}
