+++ EssentialFeediOS/Feed UI/Composer/FeedUIComposer.swift
@@ -24,39 +24,6 @@ public final class FeedUIComposer {
     }
 }
 
-private final class DispatchQueueDecorator<T> {
-    private let decoratee: T
-    init(decoratee: T) {
-        self.decoratee = decoratee
-    }
-    private func dispatch(_ completion: @escaping () -> Void) {
-        guard Thread.isMainThread else {
-            return DispatchQueue.main.async(execute: completion)
-        }
-        completion()
-    }
-}
-
-extension DispatchQueueDecorator: FeedLoader where T == FeedLoader {
-    func load(completion: @escaping (FeedLoader.Result) -> Void) {
-        decoratee.load { [weak self] result in
-            self?.dispatch {
-                completion(result)
-            }
-        }
-    }
-}
-
-extension DispatchQueueDecorator: FeedImageDataLoader where T == FeedImageDataLoader {
-    func loadImageData(from url: URL, completion: @escaping (FeedImageDataLoader.Result) -> Void) -> FeedImageDataLoaderTask {
-        return decoratee.loadImageData(from: url) { [weak self] result in
-            self?.dispatch {
-                completion(result)
-            }
-        }
-    }
-}
-
 private extension FeedViewController {
     static func makeWith(delegate: FeedViewControllerDelegate, title: String) -> FeedViewController {
         let storyBoard = UIStoryboard(name: "Feed", bundle: Bundle(for: FeedViewController.self))
